Supports passing colors directly by atom
Previously when defining elements, it was necessary to look up the
integer constant for colors and style attributes when setting them.
For example, the constant for the color red had to be looked up by
calling `Constants.color(:red)`.

The renderer now does this lookup automatically when it sees an atom.
For backwards compatibility, integer constants can also be passed as
before.