Support multi-line content in labels
When a line feed character (\n) is encountered in the content string,
the label will continue on the next line. Note that text elements
are inline and don't support multiline content.

Automatic line wrapping is supported via `wrap` attribute, but note
that it doesn't reflow existing line breaks, as that would require
understanding the content's structure. It's best to use the `wrap`
attribute on content without line breaks already in place.

Closes #8